move login para o index e sidebar em componente
