+++ client/src/components/Sidebar.tsx
@@ -142,11 +142,6 @@ const Sidebar = () => {
           ))}
         </ul>
       </div>
-
-      {/* // Page Content */}
-      <div className="p-7">
-        <h1 className="text-2xl font-semibold text-lightGreen">SlideBar</h1>
-      </div>
     </div>
   );
 };
