--- client/src/pages/Register.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Register = () => {
-  return (
-    <div>Register</div>
-  )
-}
-
-export default Register
